Move SelectComponents under hoc folder
