+++ packages/react-vapor/src/components/select/hoc/tests/MultiSelectWithPredicate.spec.tsx
@@ -18,8 +18,8 @@ import {IItemBoxProps} from '../../../itemBox/ItemBox';
 import {reorderListBoxOption, unselectListBoxOption} from '../../../listBox/ListBoxActions';
 import {IMultiSelectOwnProps, IMultiSelectProps, MultiSelectConnected} from '../../MultiSelectConnected';
 import {toggleSelect} from '../../SelectActions';
-import {MultiSelectWithPredicate} from '../../SelectComponents';
 import {SelectConnected} from '../../SelectConnected';
+import {MultiSelectWithPredicate} from '../SelectComponents';
 import {ISelectWithPredicateProps, selectWithPredicate} from '../SelectWithPredicate';
 
 describe('Select', () => {
